+++ src/main/kotlin/me/rafaelldi/aspire/actions/OpenAspireDashboardAction.kt
@@ -1,24 +1,43 @@
 package me.rafaelldi.aspire.actions
 
-import com.intellij.icons.AllIcons
 import com.intellij.ide.BrowserUtil
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import me.rafaelldi.aspire.AspireBundle
+import me.rafaelldi.aspire.sessionHost.AspireSessionHostManager
+import me.rafaelldi.aspire.util.SESSION_HOST_ID
 
-class OpenAspireDashboardAction(private val dashboardUrl: String?) : AnAction(
-    AspireBundle.message("action.Aspire.Dashboard.Open.text"),
-    AspireBundle.message("action.Aspire.Dashboard.Open.description"),
-    AllIcons.General.Web
-) {
+class OpenAspireDashboardAction : AnAction() {
     override fun actionPerformed(event: AnActionEvent) {
-        if (dashboardUrl == null) return
+        val project = event.project ?: return
+        val sessionHostId = event.getData(SESSION_HOST_ID) ?: return
+        val sessionHost = AspireSessionHostManager
+            .getInstance(project)
+            .getSessionHost(sessionHostId)
+            ?: return
+        val dashboardUrl = sessionHost.hostData.dashboardUrl
+        if (dashboardUrl.isNullOrEmpty()) return
+
         BrowserUtil.browse(dashboardUrl)
     }
 
     override fun update(event: AnActionEvent) {
-        event.presentation.isEnabled = dashboardUrl != null
+        val project = event.project
+        val sessionHostId = event.getData(SESSION_HOST_ID)
+        if (project == null || sessionHostId == null) {
+            event.presentation.isEnabledAndVisible = false
+            return
+        }
+
+        val hostAvailable = AspireSessionHostManager
+            .getInstance(project)
+            .isSessionHostAvailable(sessionHostId)
+        if (!hostAvailable) {
+            event.presentation.isEnabledAndVisible = false
+            return
+        }
+
+        event.presentation.isEnabledAndVisible = true
     }
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT