+++ src/main/kotlin/me/rafaelldi/aspire/run/AspireHostProgramRunner.kt
@@ -24,7 +24,6 @@ class AspireHostProgramRunner : DotNetProgramRunner() {
     companion object {
         const val DEBUG_SESSION_TOKEN = "DEBUG_SESSION_TOKEN"
         const val DEBUG_SESSION_PORT = "DEBUG_SESSION_PORT"
-        const val ASPNETCORE_URLS = "ASPNETCORE_URLS"
         const val DOTNET_DASHBOARD_OTLP_ENDPOINT_URL = "DOTNET_DASHBOARD_OTLP_ENDPOINT_URL"
         private const val RUNNER_ID = "aspire-runner"
 
@@ -50,13 +49,15 @@ class AspireHostProgramRunner : DotNetProgramRunner() {
             throw CantRunException("Unable to find token or port")
         LOG.trace("Found $DEBUG_SESSION_TOKEN $debugSessionToken and $DEBUG_SESSION_PORT $debugSessionPort")
 
-        val dashboardUrl = environmentVariables[ASPNETCORE_URLS]
-        LOG.trace("Found $ASPNETCORE_URLS $dashboardUrl")
         val openTelemetryProtocolUrl = environmentVariables[DOTNET_DASHBOARD_OTLP_ENDPOINT_URL]
         LOG.trace("Found $DOTNET_DASHBOARD_OTLP_ENDPOINT_URL $openTelemetryProtocolUrl")
 
         val runProfileName = environment.runProfile.name
         val isDebug = environment.executor.id == DefaultDebugExecutor.EXECUTOR_ID
+
+        val parameters =
+            (environment.runnerAndConfigurationSettings?.configuration as? AspireHostConfiguration)?.parameters
+        val dashboardUrl = parameters?.startBrowserParameters?.url
 
         val aspireHostLifetime = environment.project.lifetime.createNested()